requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
requests
ddgs
pydantic
streamlit